perfdata: Review util funcs
